Corrette funzioni makeTime(), fromStrToTimeEl(), cambiato tipo result del parser da float a double (numeri a 10 cifre per tempo UTC)

+++ Software/RollerShutter/abparser.h
@@ -3,8 +3,8 @@
 #include <Arduino.h>
 #include "common.h"
 
-extern float variables(char *);
-extern float actions(char *, float);
-float eval(const char *str);
+extern double variables(char *);
+extern double actions(char *, double);
+double eval(const char *str);
 
 #endif //__ABPARSER__
